AI Generated Image Detection/requirements.txt
torch
opencv-python
Pillow
pandas
tqdm
scikit-learn
numpy
einops
torchvision
timm
scikit-image